+ pybuild: initial checkin, builds static python-12.4, modules builtin
this took a while to get right, but there are still problems, notably,
loadable extension modules do not seem to compile and we can't figure
this out at the moment.  wheels do install however, and even seem to
work ok, but they likely will cause a second copy of libc (ie, from
libc.so) to be loaded.

we might just switch to a dynamic build (with static libssl/libcrypto),
one per system and rebuild them all each release (of openssl or python),
while we figure out if it's even possible to use a statically linked
python with compiled loadable extension modules.

it seems that not many are doing this, but musl (which links statically)
might be a hint, although it seems that they disable loadable modules,
perhaps because it doesn't work (this may be due to dlopen() not working
from static executables).

more info describing [at least some elements of] the problem can be
found at README of https://github.com/pfalcon/foreign-dlopen project