/ pybuild: dynamic: need to remove explicit _typing on dynamic
this works fine on static, but for some reason this has to be removed
for dynamic build.  either way it gets included (determined by test
"import _typing"), it's just that we can't specify it in Setup.local or
it has build problems.  this is just a faster C implementation of
"typing" that was added recently (3.10? 3.11 iirc?) so maybe some aspect
of build machinery has not caught up yet... in any case it appears to
build and work fine without specifying it.