/ pybuild: setup: sslinc: include dir should not have trailing openssl/
this must not have been required to work because the files use the
include idiom like this:

  #include <openssl/foo.h>

so it would have checked openssl/openssl and not found it.

my guess is, earlier successful builds used the headers from the system
include dirs after not finding the ones we supplied with -I and
incorrect path.  but when the linker did its work, it correctly used the
lib from /opt/ (we checked with "import ssl; ssl.OPENSSL_VERSION").
this happened to work before because we have the same major and minor
version in our own build as the system.  it failed when we built against
it on an older linux, and then we noticed the error.