+ pylinks: new script, makes symlinks to run python from different root
typically we install python in /opt/ but don't want to put that dir in
our path.  instead we make symlinks in /usr/local/.  this automates that
process with the full set of links.  we can't only put bin/ links, we
have to put all of them so that python3-config works correctly, for
example.  it returns relative locations depending only on the invocation
path.