- pylinks: make absolute links on systems without "ln -r"
check the binary help output, older coreutils doesn't have the ln
'--relative' flag.  we could make the link ourself using realpath, but
realpath itself is not present on the oldest systems, and in reality,
these systems without the flag won't exist for much longer.  we'll just
fall back to absolute links anyways, which aren't as good, but not
horrible.