+ makefile: readme: stub recipe to write auto generator
it will read the comment header in all the executables and generate a
readme.rst from them, automatically.  this is already the origin of
the present readme (about to commit).

contains the basic recipe in a comment, that does the scrape, everything
could be added later with pure text transforms.  we will implement it
later, for now have to get a readme done