+ dnsroute: new tool maintains static route to an ip determined by dns
this allows for it to move.  uses a statefile, but there's really no
other way to do it, because we wouldn't have any way to know which route
entry to delete after it changed, so they would pile up.  we also
support the gateway changing, but this is thought to be atypical.  the
use case for now is smtp-relay@gmail.com which seems to change quite a
bit and looks up differently in different regions.  we'll need to run
this out of cron on affected systems.