+ kubectlpty: wrapper to invoke kubectl via pty for kubecolor
need to fool kubectl it into think it's invoked from a terminal so it
does width-wrapping.  otherwise all lines are infinite length.  see
kubecolor issue 279.

don't like that the binary is in utiladm but this is kind of unavoidable
in case of C programs, they aren'c convenient like scripts.  at the
moment we aren't running kubectl on any non-amd64 system but obviously
this would not work right if we needed it on arm64.  and binaries in git
repo, while it can be ok for one-off, is not a good practice.  have to
think about this more...