/ kctl: args: were not passed properly, artifact of devskel
when we switched from using function in bash profile we directly
processed args, but this has since moved to a script based on our
devskel.  although this is not really using any actual part of devskel,
we support the devskel names and invocation format so we can add to it
later.  but we forgot to adapt the script and actually add the
arguments, which normally get processed by process_args() but we don't
use that.  nonetheless we need to rename $@ as $args.