/ kctl: missed spot to quote expanded variable in previous commit
long time ago we learned that these args do not need to be quoted in
double bracket expression, they can happily have spaces, etc.  but since
we're debugging a completions issue we'll be ultra careful and quote
everything.  this one tacks on to a previous commit, we just missed one
location to quote.