/ kctl: do not use kubecolor when invoked indirectly for completions
seems to be some bug if we use kubecolor for completions, the output
gets mangled and has carriage returns on it and some weird stray
characters.  first thought kubecolor was responsible but it actually
works fine when not going through the script, using alias kctl=kubecolor
for example, completions work fine.  so, it's not clear where the bug
lies.  tried for a while to figure it out but ran out of time and
figured it's not worth it at the moment we just don't have time to work
on that.

at first we bypassed the script entirely and did an "exec kubectl" at
script entrance if $1 was __complete, but decided to do the processing
because that way we can get proper completion using the namespaces/args
changes that the script implements.  since it seems to work, may as well
keep that stuff and completions should better correspond to what we're
intending with the wrapper.