+ kctl: let the user override namespace set by kns on per-command basis
if given before the command word (ie as one of the initial options) then
let the user override, on a per-invocation basis, the namespace he has
selected in $KUBENS ie with kns() bash function, which sets KUBENS (and
it was set for export during shell init already).

this wrapper will end up as a pile of hacks... sad to say.